Architectures and Partitions
Beginings of the meat of the problem, here we define models for architectures
and partition tables. For each resource type probe the API for existing resources
adding defined ones if not present and also purging things added in automatically
by the seed process (i.e. delete stuff not in version control).