Genericize Resource Management
Resource types mixin a base model type.  This is responsible for performing definition
of whitelisted attributes at the class level, instance initialization based on the
whitelisted items.  The main algorithm is also abstracted away to be generic in that
it will now probe for all existing instances of a resource type, create new instances
defined in the configuration, perform any updates to the attributes, then lazily commit
results back and also purge unmanaged resources.