Clean Up Backend Code
In order to genericize backend API code each resource type now defines its path
as a class instance variable, thus the all and commit methods are looking a lot
more like one another.  To clean up the JSON payload name has been rolled into
an attribute and the attributes array can just be passed in directly, this leads
to a mismatch between the configuration and internal representation however which
may need to be considered later on.