Separate Out Shared Collector Code
There are two classes of 'collectors' which implmenent the 'all' class method for each
resource type.  The basic one is able to define all resources with a single GET to the
API for that resource.  A deep version is required for things like partition tables and
templates whose content is only returned when performing a GET on the API for that
specific resource.  This functionality has been shifted into a mixin so it can be reused
by new resource types and shared.