Operating System Support
Introduces the new operating system model.  This extends the base model to allow attributes
to be specified multiple times, accumulating as it goes.  The collectors will also now look
for a class method called munge which will take what Foreman offers us and allows us to hook
in an modify the input.  In this case we need to create ID arrays used in POST/PUT commands
and also expected by the class attributes.  Additionally we add in an abstraction for resource
title, which defaults to name, but in the case of operating systems allows a unique composite
to be created. A title in the configuration now overrides the name when checking for matching
existing resources.  Configuration is now deep processed for file and resource references
as checking just top level values wasn't deep enough, we also sort all arrays for comparison
operations to succeed.